+++ app/smartbrain_3000/web/_app/immutable/entry/start.pCvaIQeJ.js
@@ -1,1 +1,0 @@
-import{a as e,n as t}from"../chunks/BaD4UxYV.js";export{e as load_css,t as start};